+++ src/libtropic.c
@@ -15,6 +15,7 @@
 
 #include "TROPIC01_configuration_objects.h"
 #include "libtropic_common.h"
+#include "libtropic_logging.h"
 #include "libtropic_port.h"
 #include "lt_aesgcm.h"
 #include "lt_asn1_der.h"
@@ -1303,7 +1304,7 @@ const char *lt_ret_verbose(lt_ret_t ret)
 
 /** @brief This helper structure together with two get* interfaces is meant to be used to simplify looping
  *         through all config addresses and printing them out into log */
-struct lt_config_obj_desc_t config_description_table[27] = {
+struct lt_config_obj_desc_t config_description_table[LT_CONFIG_OBJ_CNT] = {
     {"CONFIGURATION_OBJECTS_CFG_START_UP                   ", CONFIGURATION_OBJECTS_CFG_START_UP_ADDR},
     {"CONFIGURATION_OBJECTS_CFG_SLEEP_MODE                 ", CONFIGURATION_OBJECTS_CFG_SLEEP_MODE_ADDR},
     {"CONFIGURATION_OBJECTS_CFG_SENSORS                    ", CONFIGURATION_OBJECTS_CFG_SENSORS_ADDR},
@@ -1333,15 +1334,27 @@ struct lt_config_obj_desc_t config_description_table[27] = {
     {"CONFIGURATION_OBJECTS_CFG_UAP_MAC_AND_DESTROY        ", CONFIGURATION_OBJECTS_CFG_UAP_MAC_AND_DESTROY_ADDR},
     {"CONFIGURATION_OBJECTS_CFG_UAP_SERIAL_CODE_GET        ", CONFIGURATION_OBJECTS_CFG_UAP_SERIAL_CODE_GET_ADDR}};
 
-uint16_t get_conf_addr(uint8_t i) { return config_description_table[i].addr; }
+uint16_t get_conf_addr(uint8_t i)
+{
+    LT_ASSERT(1, i < LT_CONFIG_OBJ_CNT);
+    return config_description_table[i].addr;
+}
 
-const char *get_conf_desc(uint8_t i) { return config_description_table[i].desc; }
+const char *get_conf_desc(uint8_t i)
+{
+    LT_ASSERT(1, i < LT_CONFIG_OBJ_CNT);
+    return config_description_table[i].desc;
+}
 
 lt_ret_t read_whole_R_config(lt_handle_t *h, struct lt_config_t *config)
 {
+    if (!h || !config) {
+        return LT_PARAM_ERR;
+    }
+
     lt_ret_t ret;
 
-    for (uint8_t i = 0; i < 27; i++) {
+    for (uint8_t i = 0; i < LT_CONFIG_OBJ_CNT; i++) {
         ret = lt_r_config_read(h, get_conf_addr(i), &config->obj[i]);
         if (ret != LT_OK) {
             return ret;
@@ -1353,9 +1366,13 @@ lt_ret_t read_whole_R_config(lt_handle_t *h, struct lt_config_t *config)
 
 lt_ret_t write_whole_R_config(lt_handle_t *h, const struct lt_config_t *config)
 {
+    if (!h || !config) {
+        return LT_PARAM_ERR;
+    }
+
     lt_ret_t ret;
 
-    for (int i = 0; i < 27; i++) {
+    for (int i = 0; i < LT_CONFIG_OBJ_CNT; i++) {
         ret = lt_r_config_write(h, get_conf_addr(i), config->obj[i]);
         if (ret != LT_OK) {
             return ret;
@@ -1367,9 +1384,13 @@ lt_ret_t write_whole_R_config(lt_handle_t *h, const struct lt_config_t *config)
 
 lt_ret_t read_whole_I_config(lt_handle_t *h, struct lt_config_t *config)
 {
+    if (!h || !config) {
+        return LT_PARAM_ERR;
+    }
+
     lt_ret_t ret;
 
-    for (uint8_t i = 0; i < 27; i++) {
+    for (uint8_t i = 0; i < LT_CONFIG_OBJ_CNT; i++) {
         ret = lt_i_config_read(h, get_conf_addr(i), &config->obj[i]);
         if (ret != LT_OK) {
             return ret;
@@ -1381,6 +1402,10 @@ lt_ret_t read_whole_I_config(lt_handle_t *h, struct lt_config_t *config)
 
 lt_ret_t verify_chip_and_start_secure_session(lt_handle_t *h, uint8_t *shipriv, uint8_t *shipub, uint8_t pkey_index)
 {
+    if (!h || !shipriv || !shipub || (pkey_index > PAIRING_KEY_SLOT_INDEX_3)) {
+        return LT_PARAM_ERR;
+    }
+
     lt_ret_t ret = LT_FAIL;
 
     // This is not used here, but let's read it anyway